move jitclasses -> search
