+++ alphadia/raw_data/__init__.py
@@ -5,8 +5,10 @@ from alphadia.raw_data.alpharaw_wrapper import (
     Thermo,
 )
 from alphadia.raw_data.bruker import TimsTOFTranspose
-from alphadia.raw_data.jitclasses.alpharaw_jit import AlphaRawJIT
-from alphadia.raw_data.jitclasses.bruker_jit import TimsTOFTransposeJIT
+
+# TODO: inverse the dependency here:
+from alphadia.search.jitclasses.alpharaw_jit import AlphaRawJIT
+from alphadia.search.jitclasses.bruker_jit import TimsTOFTransposeJIT
 
 DiaData = TimsTOFTranspose | AlphaRawBase | MzML | Sciex | Thermo
 DiaDataJIT = TimsTOFTransposeJIT | AlphaRawJIT
